@@ -1,16 +1,15 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\..\fsc.props" />
-  <Import Project="..\..\netfx.props" />
   <PropertyGroup>
-    <TargetFrameworks>net461;netcoreapp2.0</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1</TargetFrameworks>
+    <!--<RuntimeIdentifiers>win</RuntimeIdentifiers>-->
     <AssemblyName>SqlClient.Tests</AssemblyName>
     <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <DefineConstants>$(DefineConstants);WITH_LEGACY_NAMESPACE</DefineConstants>
     <NoWarn>101</NoWarn>
-    <!-- 
-    disable error FS0044: This construct is deprecated. use 'FSharp.Data.SqlClient.Internals.ISqlCommand' instead 
+    <!--
+    disable error FS0044: This construct is deprecated. use 'FSharp.Data.SqlClient.Internals.ISqlCommand' instead
     for the time necessary of the transition, this enables more coverage of the type aliases working the way expected in client code
     -->
     <NoWarn>$(NoWarn);0044</NoWarn>
@@ -53,7 +52,7 @@
   </ItemGroup>
   <ItemGroup>
     <Reference Include="FSharp.Data.SqlClient">
-      <HintPath>..\..\bin\netstandard2.0\FSharp.Data.SqlClient.dll</HintPath>
+      <HintPath>..\..\bin\netstandard2.1\FSharp.Data.SqlClient.dll</HintPath>
     </Reference>
   </ItemGroup>
   <Target Name="CopyCustomContent" AfterTargets="AfterBuild">
