@@ -1,8 +1,8 @@
 ﻿namespace FSharp.Data.SqlClient
 
-open System.Configuration
-open System.IO
 open System
+open System.IO
+open System.Configuration
 open System.Collections.Generic
 open System.Diagnostics
 
@@ -47,7 +47,7 @@ type internal DesignTimeConnectionString =
         match configSection, lazy configSection.[name] with
         | null, _ | _, Lazy null -> raise <| KeyNotFoundException(message = sprintf "Cannot find name %s in <connectionStrings> section of %s file." name configFilename)
         | _, Lazy x -> 
-            let providerName = if String.IsNullOrEmpty x.ProviderName then "System.Data.SqlClient" else x.ProviderName
+            let providerName = if String.IsNullOrEmpty x.ProviderName then "Microsoft.Data.SqlClient" else x.ProviderName
             x.ConnectionString, providerName
 
     member this.Value = 
