@@ -2,10 +2,8 @@
 
 open System
 open System.Data
-open System.Data.SqlClient
+open Microsoft.Data.SqlClient
 open System.Reflection
-open System.Collections
-open System.Diagnostics
 
 open Microsoft.FSharp.Quotations
 
