@@ -6,9 +6,9 @@ open System.Data
 open System.Collections.Generic
 open System.IO
 open System.Threading.Tasks
-open System.Data.SqlClient
+
+open Microsoft.Data.SqlClient
 open FSharp.Data.SqlClient.Internals
-open System.Diagnostics
 
 type internal TypeInfoPerConnectionStringCache() =
   let key = obj()
