@@ -3,7 +3,7 @@
 open System
 open System.Data
 open System.Transactions
-open System.Data.SqlClient
+open Microsoft.Data.SqlClient
 
 open Xunit
 
