@@ -1,13 +1,13 @@
 #if WITH_LEGACY_NAMESPACE
 module FSharp.Data.TypeProviderTest
-open FSharp.Data.SqlClient
 #else
 module FSharp.Data.SqlClient.TypeProviderTest
 #endif
 
 open System
 open System.Data
-open System.Data.SqlClient
+open Microsoft.Data.SqlClient
+open FSharp.Data.SqlClient
 open Xunit
 
 type GetEvenNumbers = SqlCommandProvider<"select * from (values (2), (4), (8), (24)) as T(value)", ConnectionStrings.AdventureWorksNamed>
@@ -106,7 +106,7 @@ let ToTraceString() =
 let runScalarQuery query = 
     use conn = new SqlConnection(ConnectionStrings.AdventureWorks)
     conn.Open()
-    use cmd = new System.Data.SqlClient.SqlCommand()
+    use cmd = new Microsoft.Data.SqlClient.SqlCommand()
     cmd.Connection <- conn  
     cmd.CommandText <- query
     cmd.ExecuteScalar()
